+++ tpl/js/validator.js
@@ -1,6 +1,6 @@
 ;(function(app){
 
-	app.editor.validator('name of this rule', function(options, val, form){
+	app.editor.validator('{{{name}}}', function(options, val, form){
 
 		//return nothing if ok
 		//return error message 'string' if not